requirements.txt
streamlit>=1.28.0
transformers>=4.35.0
pandas>=2.0.3
pdfplumber>=0.10.3
pytesseract>=0.3.10
Pillow>=10.1.0
python-dotenv>=1.0.0
yfinance>=0.2.31
prophet>=1.1.3
plotly>=5.18.0
scikit-learn>=1.3.0
nltk>=3.8.1
requests==2.31.0
newspaper3k==0.2.8